Refactor Navbar into Modular Components with Improved Code Organization
- Split Navbar into smaller, focused components (Logo, DarkModeToggle, UserMenu, etc.)
- Simplified main Navbar logic by extracting complex UI elements
- Removed LanguageSwitcher component from the project
- Updated import paths and restructured component hierarchy
- Improved code readability and maintainability
- Preserved existing functionality and styling
