+++ src/components/layout/Navbar.jsx
@@ -1,14 +1,18 @@
-// src/components/layout/Navbar.jsx
-import { useState, useEffect, useRef } from 'react';
-import { Link, useLocation, useNavigate } from 'react-router-dom';
-import { motion, AnimatePresence } from 'framer-motion';
-import { useLanguage } from '../../contexts/LanguageContext';
-import { useDarkMode } from '../../contexts/DarkModeContext';
+// src/components/layout/Navbar/index.jsx
+import { useState, useEffect } from 'react';
+import { useLocation } from 'react-router-dom';
 import { auth, db } from '../../config/firebase';
-import { signOut, onAuthStateChanged } from 'firebase/auth';
+import { onAuthStateChanged } from 'firebase/auth';
 import { doc, getDoc } from 'firebase/firestore';
-import { MoonIcon, SunIcon, MenuIcon, XIcon, ChevronDownIcon } from 'lucide-react';
-import LanguageSwitcher from '../common/LanguageSwitcher'; // Sesuaikan path jika diperlukan
+import { useDarkMode } from '../../../contexts/DarkModeContext';
+import { useLanguage } from '../../../contexts/LanguageContext';
+
+import Logo from './Logo';
+import DarkModeToggle from './DarkModeToggle';
+import DesktopNav from './DesktopNav';
+import UserMenu from './UserMenu';
+import MobileMenu from './MobileMenu';
+import AuthButtons from './AuthButtons';
 
 // Simple translations object
 const translations = {
@@ -40,17 +44,14 @@ const translations = {
   }
 };
 
-export default function Navbar({ onMenuClick, isLoggedIn = false }) {
+export default function Navbar({ onMenuClick }) {
   const [userData, setUserData] = useState(null);
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
   const [isScrolled, setIsScrolled] = useState(false);
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
   const [isAuthCheckComplete, setIsAuthCheckComplete] = useState(false);
-  const dropdownRef = useRef(null);
   const location = useLocation();
-  const navigate = useNavigate();
-  const { language, toggleLanguage } = useLanguage();
-  const { isDarkMode, toggleDarkMode } = useDarkMode();
+  const { language } = useLanguage();
+  const { isDarkMode } = useDarkMode();
 
   // Get translations based on current language
   const t = translations[language] || translations.en;
@@ -108,17 +109,6 @@ export default function Navbar({ onMenuClick, isLoggedIn = false }) {
     return () => window.removeEventListener('scroll', handleScroll);
   }, []);
 
-  // Close dropdown when clicking outside
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (dropdownRef.current && !dropdownRef.current.contains(event.target)) {
-        setIsDropdownOpen(false);
-      }
-    };
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => document.removeEventListener('mousedown', handleClickOutside);
-  }, [isDropdownOpen]);
-
   // Close mobile menu on route change
   useEffect(() => {
     setIsMobileMenuOpen(false);
@@ -129,40 +119,6 @@ export default function Navbar({ onMenuClick, isLoggedIn = false }) {
     ? isScrolled ? 'bg-gray-900/80 border-b border-primary-900/30' : 'bg-transparent'
     : isScrolled ? 'bg-white/80 border-b border-primary-200/40 shadow-lg shadow-primary-500/5' : 'bg-transparent';
 
-  // Consistent text colors across both modes
-  const navLinkColor = isDarkMode
-  ? 'text-primary-400 hover:text-primary-300'
-  : isScrolled 
-    ? 'text-primary-600 hover:text-primary-700' 
-    : 'text-primary-600 hover:text-primary-700'; 
-  
-  // Logo background effect for dark/light mode  
-  const logoBgEffect = isDarkMode 
-    ? 'from-primary-700/20 via-secondary-700/20 to-primary-700/20'
-    : 'from-primary-500/20 via-secondary-500/20 to-primary-500/20';
-
-  // Active state background
-  const activeNavBg = isDarkMode
-    ? 'bg-primary-900/30'
-    : isScrolled ? 'bg-primary-50/70' : 'bg-white/20';
-
-  // Handle sign out
-  const handleSignOut = async () => {
-    try {
-      await signOut(auth);
-      // Redirect to home page after signing out
-      navigate('/');
-    } catch (error) {
-      console.error("Error signing out: ", error);
-    }
-  };
-
-  // Get user initials for avatar
-  const getUserInitials = (name) => {
-    if (!name) return "U";
-    return name.split(' ').map(n => n[0]).join('').toUpperCase();
-  };
-
   // Nav links data
   const navLinks = [
     { name: t.features, path: '/features' },
@@ -171,6 +127,7 @@ export default function Navbar({ onMenuClick, isLoggedIn = false }) {
     { name: t.blog, path: '/blog' }
   ];
 
+  const dashboardLink = { name: t.dashboard, path: '/dashboard' };
   const isUserLoggedIn = !!userData;
 
   return (
@@ -191,401 +148,79 @@ export default function Navbar({ onMenuClick, isLoggedIn = false }) {
         <div className="flex justify-between h-16 sm:h-20 items-center">
           {/* Left side with logo and menu button */}
           <div className="flex items-center space-x-4">
-            {/* Dashboard menu button (only for logged in users) */}
-            {isUserLoggedIn && (
-              <motion.button
-                onClick={onMenuClick}
-                className={`lg:hidden relative p-2 rounded-lg text-primary-500 transition-all duration-300 hover:text-primary-400`}
-                whileHover={{ scale: 1.1 }}
-                whileTap={{ scale: 0.9 }}
-                aria-label="Open main menu"
-              >
-                <span className={`absolute inset-0 rounded-lg opacity-0 hover:opacity-100 ${
-                  isDarkMode ? 'bg-primary-900/50' : 'bg-primary-50/70'
-                } transition-opacity duration-300`}></span>
-                <MenuIcon size={20} />
-              </motion.button>
-            )}
-            
-            {/* Mobile menu button (for non-logged in users) */}
-            {!isUserLoggedIn && (
-              <motion.button 
+            {/* Menu buttons */}
+            {isUserLoggedIn ? (
+              <MobileMenu.DashboardButton onMenuClick={onMenuClick} isDarkMode={isDarkMode} />
+            ) : (
+              <MobileMenu.Button 
+                isOpen={isMobileMenuOpen}
                 onClick={() => setIsMobileMenuOpen(!isMobileMenuOpen)}
-                className={`md:hidden relative p-2 rounded-lg text-primary-500 transition-all duration-300 hover:text-primary-400`}
-                whileHover={{ scale: 1.1 }}
-                whileTap={{ scale: 0.9 }}
-                aria-label="Toggle mobile menu"
-              >
-                <span className={`absolute inset-0 rounded-lg opacity-0 hover:opacity-100 ${
-                  isDarkMode ? 'bg-primary-900/50' : 'bg-primary-50/70'
-                } transition-opacity duration-300`}></span>
-                {isMobileMenuOpen ? <XIcon size={20} /> : <MenuIcon size={20} />}
-              </motion.button>
+                isDarkMode={isDarkMode}
+              />
             )}
             
-            {/* Logo with animation effects */}
-            <Link to="/" className="flex-shrink-0 group relative">
-              <div className={`absolute -inset-4 rounded-2xl bg-gradient-to-r ${logoBgEffect} opacity-0 blur-xl group-hover:opacity-100 group-hover:duration-1000 duration-300 group-hover:animate-pulse-slow`}></div>
-              <div className="relative flex items-center">
-                <div className="relative">
-                  <div className="font-heading font-bold text-2xl md:text-3xl bg-gradient-to-r from-primary-500 via-secondary-400 to-primary-500 bg-clip-text text-transparent bg-size-200 bg-pos-0 group-hover:bg-pos-100 transition-all duration-500">
-                    Growthify
-                  </div>
-                  {/* Animated underline effect */}
-                  <div className="absolute left-0 bottom-0 w-full h-[2px] bg-gradient-to-r from-primary-500 via-secondary-400 to-primary-500 transform scale-x-0 group-hover:scale-x-100 transition-transform duration-500 origin-left"></div>
-                </div>
-                {/* Orbiting dot animation */}
-                <div className="absolute -right-1 -top-1 w-2 h-2 rounded-full bg-primary-500 opacity-0 group-hover:opacity-100 group-hover:animate-orbit transition-opacity duration-300"></div>
-              </div>
-            </Link>
+            {/* Logo */}
+            <Logo isDarkMode={isDarkMode} />
           </div>
           
           {/* Navigation Links - Desktop */}
           {isAuthCheckComplete && !isUserLoggedIn && (
-            <div className="hidden md:flex md:items-center md:space-x-1">
-              {navLinks.map((link, index) => {
-                const isActive = location.pathname === link.path;
-                return (
-                  <Link 
-                    key={link.name} 
-                    to={link.path} 
-                    className={`px-4 py-2 mx-1 rounded-lg text-sm font-medium relative overflow-hidden transition-all duration-300 ${
-                      isActive 
-                        ? isDarkMode ? 'text-primary-300' : 'text-primary-600' 
-                        : navLinkColor
-                    }`}
-                  >
-                    {/* Dynamic highlight effect */}
-                    <span className={`absolute inset-0 rounded-lg -z-10 transition-opacity duration-300 
-                      ${isActive ? activeNavBg : 'opacity-0'}`}
-                    ></span>
-
-                    {/* Hover highlight */}
-                    <span className={`absolute inset-0 rounded-lg -z-10 opacity-0 hover:opacity-100 transition-opacity duration-300 ${
-                      isDarkMode ? 'bg-primary-900/50' : isScrolled ? 'bg-primary-50/70' : 'bg-white/20'
-                    }`}></span>
-                    
-                    {/* Text with dot indicator for active state */}
-                    <span className="relative flex items-center">
-                      {link.name}
-                      {isActive && (
-                        <span className="absolute -bottom-1 left-1/2 transform -translate-x-1/2 w-1 h-1 rounded-full bg-primary-500"></span>
-                      )}
-                    </span>
-                    
-                    {/* Animated underline with gradient */}
-                    <span className="absolute bottom-0 left-0 right-0 h-[2px] bg-gradient-to-r from-primary-500/0 via-primary-500 to-primary-500/0 transform scale-x-0 hover:scale-x-100 transition-transform duration-300 ease-out"></span>
-                  </Link>
-                );
-              })}
-            </div>
+            <DesktopNav 
+              links={navLinks} 
+              currentPath={location.pathname}
+              isDarkMode={isDarkMode}
+              isScrolled={isScrolled}
+            />
           )}
           
           {/* Dashboard Link for logged in users */}
           {isAuthCheckComplete && isUserLoggedIn && (
-            <div className="hidden md:flex md:items-center md:space-x-1">
-              <Link 
-                to="/dashboard" 
-                className={`px-4 py-2 mx-1 rounded-lg text-sm font-medium relative overflow-hidden transition-all duration-300 ${
-                  location.pathname === '/dashboard' 
-                    ? isDarkMode ? 'text-primary-300' : 'text-primary-600' 
-                    : navLinkColor
-                }`}
-              >
-                {/* Dynamic highlight effect */}
-                <span className={`absolute inset-0 rounded-lg -z-10 transition-opacity duration-300 
-                  ${location.pathname === '/dashboard' ? activeNavBg : 'opacity-0'}`}
-                ></span>
-
-                {/* Hover highlight */}
-                <span className={`absolute inset-0 rounded-lg -z-10 opacity-0 hover:opacity-100 transition-opacity duration-300 ${
-                  isDarkMode ? 'bg-primary-900/50' : isScrolled ? 'bg-primary-50/70' : 'bg-white/20'
-                }`}></span>
-                
-                {/* Text with dot indicator for active state */}
-                <span className="relative flex items-center">
-                  {t.dashboard}
-                  {location.pathname === '/dashboard' && (
-                    <span className="absolute -bottom-1 left-1/2 transform -translate-x-1/2 w-1 h-1 rounded-full bg-primary-500"></span>
-                  )}
-                </span>
-                
-                {/* Animated underline with gradient */}
-                <span className="absolute bottom-0 left-0 right-0 h-[2px] bg-gradient-to-r from-primary-500/0 via-primary-500 to-primary-500/0 transform scale-x-0 hover:scale-x-100 transition-transform duration-300 ease-out"></span>
-              </Link>
-            </div>
+            <DesktopNav 
+              links={[dashboardLink]} 
+              currentPath={location.pathname}
+              isDarkMode={isDarkMode}
+              isScrolled={isScrolled}
+            />
           )}
           
           {/* Right Side */}
           <div className="flex items-center space-x-4">
-            {/* Dark Mode Toggle with elegant effects */}
-            <motion.button
-              onClick={toggleDarkMode}
-              className="p-2 rounded-lg relative overflow-hidden text-primary-500 hover:text-primary-400 transition-colors duration-300"
-              whileHover={{ scale: 1.1 }}
-              whileTap={{ scale: 0.9 }}
-              aria-label="Toggle Dark Mode"
-            >
-              <span className={`absolute inset-0 rounded-lg opacity-0 hover:opacity-100 ${
-                isDarkMode ? 'bg-primary-900/50' : 'bg-primary-50/70'
-              } transition-opacity duration-300`}></span>
-              
-              <AnimatePresence mode="wait" initial={false}>
-                <motion.div
-                  key={isDarkMode ? 'dark' : 'light'}
-                  initial={{ opacity: 0, rotate: -30, scale: 0.5 }}
-                  animate={{ opacity: 1, rotate: 0, scale: 1 }}
-                  exit={{ opacity: 0, rotate: 30, scale: 0.5 }}
-                  transition={{ duration: 0.3 }}
-                >
-                  {isDarkMode ? <SunIcon size={18} /> : <MoonIcon size={18} />}
-                </motion.div>
-              </AnimatePresence>
-            </motion.button>
+            {/* Dark Mode Toggle */}
+            <DarkModeToggle isDarkMode={isDarkMode} />
 
-           {/* Language Switcher Component */}
-<LanguageSwitcher variant="default" />
+            {/* Language Switcher */}
+            <LanguageSwitcher variant="default" />
 
-            {/* User Menu for Logged In Users */}
+            {/* User Menu or Auth Buttons */}
             {isAuthCheckComplete && isUserLoggedIn ? (
-              <div className="relative user-dropdown" ref={dropdownRef}>
-                <motion.button 
-                  onClick={() => setIsDropdownOpen(!isDropdownOpen)}
-                  className="flex items-center space-x-2 rounded-lg p-1.5 relative overflow-hidden text-primary-500 hover:text-primary-400 transition-colors duration-300"
-                  whileHover={{ scale: 1.05 }}
-                  whileTap={{ scale: 0.95 }}
-                  aria-expanded={isDropdownOpen}
-                >
-                  <span className={`absolute inset-0 rounded-lg opacity-0 hover:opacity-100 ${
-                    isDarkMode ? 'bg-primary-900/50' : 'bg-primary-50/70'
-                  } transition-opacity duration-300`}></span>
-                  
-                  {/* Avatar with premium effects */}
-                  <div className="relative group">
-                    <div className="absolute -inset-1 rounded-full bg-gradient-to-r from-primary-500 via-secondary-500 to-primary-500 opacity-40 blur-sm group-hover:opacity-70 transition-opacity duration-300"></div>
-                    {userData.avatar ? (
-                      <img 
-                        src={userData.avatar} 
-                        alt={userData.name} 
-                        className="h-8 w-8 rounded-full object-cover shadow-md relative ring-2 ring-white/10 dark:ring-black/5"
-                      />
-                    ) : (
-                      <div className="h-8 w-8 rounded-full flex items-center justify-center text-white font-semibold bg-gradient-to-br from-primary-500 to-secondary-500 shadow-md relative ring-2 ring-white/10 dark:ring-black/5">
-                        {getUserInitials(userData.name)}
-                      </div>
-                    )}
-                  </div>
-                  
-                  <div className="hidden sm:flex items-center text-primary-500 hover:text-primary-400">
-                    <span className="text-sm font-medium">{userData.name.split(' ')[0]}</span>
-                    <ChevronDownIcon size={16} className="ml-1" />
-                  </div>
-                </motion.button>
-                
-                {/* User Dropdown Menu with elegant styling */}
-                <AnimatePresence>
-                  {isDropdownOpen && (
-                    <motion.div 
-                      className={`absolute right-0 mt-3 w-56 rounded-xl overflow-hidden z-10 ${
-                        isDarkMode 
-                          ? 'bg-gray-900/95 border border-primary-900/50 shadow-xl shadow-primary-900/20' 
-                          : 'bg-white/95 border border-primary-100/50 shadow-xl shadow-primary-500/10'
-                      }`}
-                      initial={{ opacity: 0, y: -10, scale: 0.95 }}
-                      animate={{ opacity: 1, y: 0, scale: 1 }}
-                      exit={{ opacity: 0, y: -10, scale: 0.95 }}
-                      transition={{ duration: 0.2 }}
-                    >
-                      {/* Top accent border with gradient */}
-                      <div className="absolute top-0 left-0 right-0 h-[2px] bg-gradient-to-r from-primary-600 via-secondary-500 to-primary-600"></div>
-                      
-                      {/* User Info */}
-                      <div className={`px-4 py-4 ${isDarkMode ? 'border-b border-gray-800' : 'border-b border-gray-100'}`}>
-                        <div className="flex items-center space-x-3">
-                          {/* Avatar with premium effects */}
-                          <div className="relative group">
-                            <div className="absolute -inset-1 rounded-full bg-gradient-to-r from-primary-500 via-secondary-500 to-primary-500 opacity-40 blur-sm"></div>
-                            {userData.avatar ? (
-                              <img 
-                                src={userData.avatar} 
-                                alt={userData.name} 
-                                className="h-10 w-10 rounded-full object-cover shadow-md relative ring-2 ring-white/10 dark:ring-black/5"
-                              />
-                            ) : (
-                              <div className="h-10 w-10 rounded-full flex items-center justify-center text-white font-medium bg-gradient-to-br from-primary-500 to-secondary-500 shadow-md relative ring-2 ring-white/10 dark:ring-black/5">
-                                {getUserInitials(userData.name)}
-                              </div>
-                            )}
-                          </div>
-                          <div>
-                            <p className={`text-sm font-medium ${isDarkMode ? 'text-white' : 'text-gray-900'}`}>{userData.name}</p>
-                            <p className="text-xs text-primary-500">{userData.role === 'admin' ? 'Admin' : 'Member'}</p>
-                          </div>
-                        </div>
-                      </div>
-                      
-                      {/* Menu Items with hover effects */}
-                      <div className="py-2">
-                        <Link to="/profile" className={`flex items-center px-4 py-2 text-sm ${
-                          isDarkMode ? 'text-gray-300 hover:bg-primary-900/30 hover:text-primary-300' : 'text-gray-700 hover:bg-primary-50/70 hover:text-primary-600'
-                        } transition-colors duration-200`}>
-                          <svg className="h-4 w-4 mr-3 text-primary-500 opacity-80" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z" />
-                          </svg>
-                          {t.profile}
-                        </Link>
-                        
-                        <Link to="/settings" className={`flex items-center px-4 py-2 text-sm ${
-                          isDarkMode ? 'text-gray-300 hover:bg-primary-900/30 hover:text-primary-300' : 'text-gray-700 hover:bg-primary-50/70 hover:text-primary-600'
-                        } transition-colors duration-200`}>
-                          <svg className="h-4 w-4 mr-3 text-primary-500 opacity-80" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z" />
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M15 12a3 3 0 11-6 0 3 3 0 016 0z" />
-                          </svg>
-                          {t.settings}
-                        </Link>
-                        
-                        <Link to="/help" className={`flex items-center px-4 py-2 text-sm ${
-                          isDarkMode ? 'text-gray-300 hover:bg-primary-900/30 hover:text-primary-300' : 'text-gray-700 hover:bg-primary-50/70 hover:text-primary-600'
-                        } transition-colors duration-200`}>
-                          <svg className="h-4 w-4 mr-3 text-primary-500 opacity-80" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M8.228 9c.549-1.165 2.03-2 3.772-2 2.21 0 4 1.343 4 3 0 1.4-1.278 2.575-3.006 2.907-.542.104-.994.54-.994 1.093m0 3h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
-                          </svg>
-                          {t.help}
-                        </Link>
-                      </div>
-                      
-                      {/* Sign Out Button with hover effect */}
-                      <div className={`py-2 ${isDarkMode ? 'border-t border-gray-800' : 'border-t border-gray-100'}`}>
-                        <button 
-                          onClick={handleSignOut}
-                          className={`w-full text-left flex items-center px-4 py-2 text-sm ${
-                            isDarkMode 
-                              ? 'text-gray-300 hover:bg-primary-900/30 hover:text-primary-300' 
-                              : 'text-gray-700 hover:bg-primary-50/70 hover:text-primary-600'
-                          } transition-colors duration-200`}
-                        >
-                          <svg className="h-4 w-4 mr-3 text-primary-500 opacity-80" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M17 16l4-4m0 0l-4-4m4 4H7m6 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h4a3 3 0 013 3v1" />
-                          </svg>
-                          {t.signOut}
-                        </button>
-                      </div>
-                    </motion.div>
-                  )}
-                </AnimatePresence>
-              </div>
+              <UserMenu userData={userData} isDarkMode={isDarkMode} menuItems={[
+                { name: t.profile, path: '/profile', icon: 'profile' },
+                { name: t.settings, path: '/settings', icon: 'settings' },
+                { name: t.help, path: '/help', icon: 'help' },
+                { name: t.signOut, action: 'signOut', icon: 'signOut' }
+              ]} />
             ) : isAuthCheckComplete && (
-              <>
-                {/* Login Button with hover effects */}
-                <Link 
-                  to="/login" 
-                  className="hidden sm:flex items-center px-4 py-2 text-sm font-medium rounded-lg relative overflow-hidden text-primary-500 hover:text-primary-400 transition-all duration-300"
-                >
-                  <span className={`absolute inset-0 rounded-lg opacity-0 hover:opacity-100 ${
-                    isDarkMode ? 'bg-primary-900/50' : 'bg-primary-50/70'
-                  } transition-opacity duration-300`}></span>
-                  <span className="relative">{t.login}</span>
-                </Link>
-                
-                {/* Get Started Button with premium effects */}
-                <Link 
-                  to="/register" 
-                  className="relative overflow-hidden group"
-                >
-                  {/* Animated glow effect */}
-                  <div className="absolute -inset-1 bg-gradient-to-r from-primary-600 via-secondary-500 to-primary-600 rounded-lg opacity-30 group-hover:opacity-60 blur-md group-hover:blur-lg transition-all duration-500 animate-gradient-shift"></div>
-                  
-                  {/* Button with hover and press effect */}
-                  <div className={`relative flex items-center px-4 py-2 text-sm font-medium rounded-lg ${
-                    isDarkMode
-                      ? 'bg-primary-700 text-white hover:bg-primary-600 hover:shadow-lg hover:shadow-primary-700/30' 
-                      : 'bg-gradient-to-r from-primary-500 to-secondary-500 text-white hover:shadow-lg hover:shadow-primary-500/30'
-                  } transition-all duration-300 group-hover:translate-y-[1px]`}>
-                    {/* Subtle inner shimmer effect */}
-                    <span className="absolute inset-0 rounded-lg overflow-hidden">
-                      <span className="absolute -inset-[100%] top-0 block transform -skew-x-12 bg-gradient-to-r from-transparent via-white/10 to-transparent group-hover:animate-shine"></span>
-                    </span>
-                    <span className="relative">{t.getStarted}</span>
-                  </div>
-                </Link>
-              </>
+              <AuthButtons 
+                loginText={t.login} 
+                registerText={t.getStarted} 
+                isDarkMode={isDarkMode} 
+              />
             )}
           </div>
         </div>
       </div>
 
-      {/* Mobile menu - Animated slide down with premium styling */}
-      <AnimatePresence>
-        {!isUserLoggedIn && isMobileMenuOpen && (
-          <motion.div 
-            className={`md:hidden ${
-              isDarkMode 
-                ? 'bg-gray-900/95 border-b border-primary-900/30' 
-                : 'bg-white/95 border-b border-primary-100/30'
-            } shadow-xl`}
-            initial={{ height: 0, opacity: 0 }}
-            animate={{ height: 'auto', opacity: 1 }}
-            exit={{ height: 0, opacity: 0 }}
-            transition={{ duration: 0.3 }}
-          >
-            <div className="px-2 pt-2 pb-3 space-y-1">
-              {navLinks.map((link, index) => {
-                const isActive = location.pathname === link.path;
-                return (
-                  <Link 
-                    key={link.name}
-                    to={link.path} 
-                    className={`block px-4 py-3 rounded-lg text-base font-medium relative ${
-                      isActive
-                        ? 'text-primary-500'
-                        : 'text-primary-600 hover:text-primary-500'
-                    } transition-colors duration-200`}
-                    onClick={() => setIsMobileMenuOpen(false)}
-                  >
-                    <span className={`absolute inset-0 rounded-lg -z-10 ${
-                      isActive
-                        ? isDarkMode ? 'bg-primary-900/30' : 'bg-primary-50/70'
-                        : 'opacity-0 hover:opacity-100 transition-opacity duration-300'
-                    } ${isDarkMode ? 'hover:bg-primary-900/30' : 'hover:bg-primary-50/70'}`}></span>
-                    <div className="flex items-center">
-                      <span>{link.name}</span>
-                      {isActive && (
-                        <span className="ml-2 w-1.5 h-1.5 rounded-full bg-primary-500"></span>
-                      )}
-                    </div>
-                  </Link>
-                );
-              })}
-              
-              {/* CTA Button for Mobile */}
-              <div className="p-2">
-                <Link 
-                  to="/register"
-                  className="relative block overflow-hidden group"
-                  onClick={() => setIsMobileMenuOpen(false)}
-                >
-                  {/* Animated glow effect */}
-                  <div className="absolute -inset-1 bg-gradient-to-r from-primary-600 via-secondary-500 to-primary-600 rounded-lg opacity-30 group-hover:opacity-60 blur-md group-hover:blur-lg transition-all duration-500 animate-gradient-shift"></div>
-                  
-                  {/* Button with hover and press effect */}
-                  <div className={`relative flex justify-center items-center px-4 py-3 text-base font-medium rounded-lg ${
-                    isDarkMode
-                      ? 'bg-primary-700 text-white hover:bg-primary-600' 
-                      : 'bg-gradient-to-r from-primary-500 to-secondary-500 text-white'
-                  } transition-all duration-300 group-hover:translate-y-[1px]`}>
-                    {/* Subtle inner shimmer effect */}
-                    <span className="absolute inset-0 rounded-lg overflow-hidden">
-                      <span className="absolute -inset-[100%] top-0 block transform -skew-x-12 bg-gradient-to-r from-transparent via-white/10 to-transparent group-hover:animate-shine"></span>
-                    </span>
-                    <span className="relative">{t.getStarted}</span>
-                  </div>
-                </Link>
-              </div>
-            </div>
-          </motion.div>
-        )}
-      </AnimatePresence>
+      {/* Mobile Menu */}
+      {!isUserLoggedIn && (
+        <MobileMenu.Panel 
+          isOpen={isMobileMenuOpen}
+          links={navLinks}
+          currentPath={location.pathname}
+          isDarkMode={isDarkMode}
+          registerText={t.getStarted}
+          onLinkClick={() => setIsMobileMenuOpen(false)}
+        />
+      )}
     </nav>
   );
 }
